requirements.txt
expiringdict==1.2.2
google-auth==2.0.1
google-cloud-storage==1.42.0
google-api-python-client==2.18.0
google-cloud-bigquery==2.31.0
google-cloud-bigquery-storage==2.6.3
google-cloud-logging==3.0.0
google-cloud-pubsub==2.9.0
pytz==2021.1
holidays==0.13
openpyxl==3.0.10
lxml==4.9.1
tzlocal<=2.1
pendulum==2.1.2
pysftp==0.2.9
pandas==1.3.5
imbox==0.9.8
imap-tools==0.57.0
dataclass_factory==2.14
environs==9.5.0